Move debug rendering into a sub-module
